requirements.txt
Django==1.11
django-cors-headers==2.2.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
